Refactor tag package to medicaldetail, change Tag class to represent a medical tag
