+++ src/main/java/seedu/medibook/logic/parser/AddCommandParser.java
@@ -28,7 +28,7 @@ import seedu.medibook.model.patient.Ic;
 import seedu.medibook.model.patient.Patient;
 import seedu.medibook.model.patient.Phone;
 import seedu.medibook.model.patient.Weight;
-import seedu.medibook.model.tag.Tag;
+import seedu.medibook.model.medicaldetail.Tag;
 
 /**
  * Parses input arguments and creates a new AddCommand object
