+++ src/main/java/seedu/medibook/logic/parser/ParserUtil.java
@@ -22,7 +22,7 @@ import seedu.medibook.model.patient.Height;
 import seedu.medibook.model.patient.Ic;
 import seedu.medibook.model.patient.Phone;
 import seedu.medibook.model.patient.Weight;
-import seedu.medibook.model.tag.Tag;
+import seedu.medibook.model.medicaldetail.Tag;
 
 /**
  * Contains utility methods used for parsing strings in the various *Parser classes.
