+++ src/main/java/seedu/medibook/storage/JsonAdaptedPatient.java
@@ -25,7 +25,7 @@ import seedu.medibook.model.patient.Patient;
 import seedu.medibook.model.patient.Phone;
 import seedu.medibook.model.patient.Record;
 import seedu.medibook.model.patient.Weight;
-import seedu.medibook.model.tag.Tag;
+import seedu.medibook.model.medicaldetail.Tag;
 
 /**
  * Jackson-friendly version of {@link Patient}.
