+++ src/main/java/seedu/medibook/ui/patientprofile/MedicalDetailsCard.java
@@ -11,7 +11,7 @@ import javafx.scene.control.ListView;
 import javafx.scene.layout.Region;
 import javafx.util.Pair;
 import seedu.medibook.model.patient.Patient;
-import seedu.medibook.model.tag.Tag;
+import seedu.medibook.model.medicaldetail.Tag;
 import seedu.medibook.ui.UiPart;
 
 /**
