+++ src/test/java/seedu/medibook/logic/parser/AddCommandParserTest.java
@@ -57,7 +57,7 @@ import seedu.medibook.model.patient.Ic;
 import seedu.medibook.model.patient.Patient;
 import seedu.medibook.model.patient.Phone;
 import seedu.medibook.model.patient.Weight;
-import seedu.medibook.model.tag.Tag;
+import seedu.medibook.model.medicaldetail.Tag;
 import seedu.medibook.testutil.PatientBuilder;
 
 public class AddCommandParserTest {
