+++ src/test/java/seedu/medibook/logic/parser/EditCommandParserTest.java
@@ -70,7 +70,7 @@ import seedu.medibook.model.patient.Height;
 import seedu.medibook.model.patient.Ic;
 import seedu.medibook.model.patient.Phone;
 import seedu.medibook.model.patient.Weight;
-import seedu.medibook.model.tag.Tag;
+import seedu.medibook.model.medicaldetail.Tag;
 import seedu.medibook.testutil.EditPatientDescriptorBuilder;
 
 public class EditCommandParserTest {
