+++ src/test/java/seedu/medibook/testutil/EditPatientDescriptorBuilder.java
@@ -15,7 +15,7 @@ import seedu.medibook.model.patient.Ic;
 import seedu.medibook.model.patient.Patient;
 import seedu.medibook.model.patient.Phone;
 import seedu.medibook.model.patient.Weight;
-import seedu.medibook.model.tag.Tag;
+import seedu.medibook.model.medicaldetail.Tag;
 
 /**
  * A utility class to help with building EditPatientDescriptor objects.
